Move test to separat file
