+++ .travis.yml
@@ -1,0 +1,11 @@
+language: python
+python:
+  - "3.5"
+  - "3.6"
+  - "3.7"
+install:
+  - pip install coverage
+script:
+  - python -m unittest discover
+  - coverage run -m unittest discover
+  - coverage report -m --include check_http_json.py
